kaoshi/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-docx
reportlab==3.6.12
python-multipart
